requirements.txt
GitPython
python-dotenv
chardet
langchain-community
meilisearch
langchain-openai
jq